done restructuring /data and updated README to contain repo structure
