+++ src/components/global/utility/SpeechSynthesizer.tsx
@@ -2,7 +2,7 @@ import { faVolumeUp } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import clsx from "clsx";
 import React from "react";
-import getSpeechSSMLStr from "../../../data/speech-ssml";
+import getSpeechSSMLStr from "../../../data/azure-speech/speech-ssml";
 import { getTokenOrRefresh } from "../../../scripts/token_util";
 import { messageToSpeechParams } from "../../../scripts/util";
 const sdk = require("microsoft-cognitiveservices-speech-sdk");
