--- src/data/prompts/generate-data.ts
@@ -1,4 +1,0 @@
-const generateDataPrompt =
-  "I am your friend and talking with you. List ten different interesting ways you can behave or react to me, and then give a reply for each approach. Finally, for each of the ten replies, tell me whether you should use that reply if you are trying to be a good friend to me. I say, \u201C{s}\u201D List ten replies to what I just said.\r\n{intermediate_examples_formatted}";
-
-export default generateDataPrompt;
