+++ src/data/routes.ts
@@ -15,20 +15,6 @@ const routes = [
   //   name: "Ask Noora",
   //   href: "/ask-noora",
   // },
-  // {
-  //   name: "Links",
-  //   dropRoutes: [
-  //     { href: "https://oval.cs.stanford.edu/", name: "Stanford OVAL" },
-  //     {
-  //       href: "https://genie.stanford.edu/",
-  //       name: "Genie",
-  //     },
-  //     {
-  //       href: "https://profiles.stanford.edu/lynn-koegel",
-  //       name: "Dr. Koegel",
-  //     },
-  //   ],
-  // },
 ];
 
 export default routes;
