+++ src/data/statement_bank/general.ts
