+++ src/scripts/generate-evaluation-prompt.ts
@@ -1,4 +1,4 @@
-import evalPrompts from "../data/prompts/noora-eval-prompt";
+import evalPrompts from "../data/prompts/noora-eval";
 
 export default function formPrompt(
   statement: string,
